+++ apps/desktop/src/app.d.ts
@@ -8,6 +8,9 @@ declare global {
 		// interface PageState {}
 		// interface Platform {}
 	}
+		declare const __VERSION__: string
+
 }
 
-export {};
+export { };
+